ch08/requirements.txt
# Chapter 8 - Multi-Agent News Pipeline Dependencies
# Install with: pip install -r requirements.txt

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Integration
requests>=2.31.0
python-dotenv>=1.0.0

# AI/ML Libraries
openai>=1.3.0
pydantic>=2.0.0

# Database
psycopg[binary]>=3.1.0

# Jupyter Support (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Progress Bars (for batch processing)
tqdm>=4.65.0

# Date/Time Handling
pytz>=2023.3
python-dateutil>=2.8.0

# Optional but Recommended
jupyterlab>=4.0.0  # Enhanced notebook interface

